Moved support interface to cleaner namespace, as it is not used for uncleaning anymore.
